requirements.txt
streamlit
torch
numpy
transformers
janus
Pillow
